backend/requirements.txt

Flask==2.3.3
Flask-Cors==4.0.0
pillow==10.0.0
numpy==1.24.3
werkzeug==2.3.7
